feat: split type definitions into categories
