+++ src/types/auth-types.ts
@@ -1,0 +1,10 @@
+import type { useAuthStore } from '~/src/stores/auth';
+
+export interface BskyLoginResponse {
+  did: string;
+  handle: string;
+  accessJwt: string;
+  refreshJwt: string;
+}
+
+export type AuthStore = ReturnType<typeof useAuthStore>;
