+++ src/types/suggestions-types.ts
@@ -1,0 +1,16 @@
+export interface SuggestedList {
+  name: string;
+  description: string;
+  uri: string;
+}
+
+export interface SuggestionItem {
+  name: string;
+  did: string;
+  description: string;
+  suggestedLists: SuggestedList[];
+}
+
+export interface RequestCounts {
+  [date: string]: number;
+}
